@@ -2,55 +2,53 @@
   "name": "teslaterm",
   "version": "0.1.0",
   "scripts": {
-    "build-dev": "tsc && webpack --mode=development",
-    "build": "tsc && webpack --mode=production",
-    "watch": "cross-var tsc -w",
+    "build-dev": "tsc --build --incremental && webpack --mode=development",
+    "build": "tsc --build --incremental && webpack --mode=production",
     "lint": "cross-var tslint -c tslint.json -p tsconfig.json",
     "start-electron": "npm run build-dev && cross-var electron .",
     "start-node": "npm run build-dev && node js/main/main_node.js",
     "package-electron": "cross-var electron-builder --linux zip --windows zip",
-    "pkg-node": "cross-var pkg --out-path=dist/node -t  node12-linux-x64,node12-windows-x64 package.json",
+    "pkg-node": "cross-var pkg --out-path=dist/node -t  node17-linux-x64,node17-windows-x64 package.json --no-native-build",
     "pkg-rename": "cross-var mv dist/node/teslaterm-linux dist/node/ud3-node && cross-var mv dist/node/teslaterm-win.exe dist/node/ud3-node.exe",
     "package-node": "npm run pkg-node && npm run pkg-rename"
   },
   "main": "js/main/main_electron.js",
   "bin": "js/main/main_node.js",
   "dependencies": {
-    "@serialport/bindings": "10.0.0",
-    "@types/fs-extra": "^9.0.7",
-    "ini": "^1.3.5",
-    "jquery": "^3.5.1",
+    "@serialport/bindings": "10.0.1",
+    "@types/fs-extra": "^9.0.13",
+    "ini": "^2.0.0",
+    "jquery": "^3.6.0",
     "jspack": "0.0.4",
-    "jszip": "^3.5.0",
-    "justgage": "^1.3.5",
-    "jzz": "^0.6.9",
-    "midi-player-js": "^2.0.13",
+    "jszip": "^3.7.1",
+    "justgage": "^1.5.1",
+    "jzz": "^1.4.9",
+    "midi-player-js": "^2.0.16",
+    "npm-check-updates": "^12.5.3",
     "rtpmidi": "^1.0.0",
-    "serialport": "10.0.0",
-    "socket.io": "^2.3.0",
-    "socket.io-client": "^2.3.0"
+    "serialport": "10.4.0",
+    "socket.io": "^4.4.1",
+    "socket.io-client": "^4.4.1"
   },
   "devDependencies": {
-    "@types/chrome": "0.0.78",
-    "@types/ini": "^1.3.30",
-    "@types/jquery": "^3.3.38",
+    "@types/ini": "^1.3.31",
+    "@types/jquery": "^3.5.14",
     "@types/microtime": "^2.1.0",
-    "@types/node": "^16.11.7",
-    "@types/serialport": "^8.0.1",
-    "@types/socket.io": "^2.1.8",
-    "@types/socket.io-client": "^1.4.33",
-    "@types/w2ui": "^1.4.32",
-    "@types/webmidi": "^2.0.4",
+    "@types/node": "^17.0.21",
+    "@types/serialport": "^8.0.2",
+    "@types/w2ui": "^1.4.33",
+    "@types/webmidi": "^2.0.6",
     "cross-var": "^1.1.0",
-    "electron": "^17.1.0",
-    "electron-builder": "^22.7.0",
-    "node": "^16.13.0",
-    "npm": "^8.5.3",
-    "pkg": "^4.4.8",
-    "ts-loader": "^7.0.5",
-    "tslint": "^5.10.0",
+    "electron": "^17.1.2",
+    "electron-builder": "^22.14.13",
+    "node": "^17.0.0",
+    "node-abi": "^3.8.0",
+    "npm": "^8.5.4",
+    "pkg": "^5.5.2",
+    "ts-loader": "^9.2.8",
+    "tslint": "^5.20.1",
     "tslint-no-circular-imports": "^0.7.0",
-    "typescript": "^3.9.3",
+    "typescript": "^4.6.2",
     "webpack": "^5.70.0",
     "webpack-cli": "^4.9.2"
   },
@@ -70,7 +68,9 @@
       "libs/*",
       "styles/*",
       "node_modules/raphael/raphael.js",
-      "node_modules/jquery/dist/jquery.min.js"
+      "node_modules/jquery/dist/jquery.min.js",
+      "node_modules/@serialport/bindings-cpp/prebuilds/linux-x64/node.napi.glibc.node",
+      "node_modules/@serialport/bindings-cpp/prebuilds/win32-x64/node.napi.node"
     ]
   }
 }
